Restructuring of view hierarchy for side menu
Main vc is now called home vc. A new root vc holds a contentViewController
that controls the displayed content. Right now the content view
controller is set to the home vc, which displays the summoner's match
history. The root vc also holds a reference to a leftMenuViewController,
which should eventually control a view with a table view that contains
the side menu selections. Tapping the side menu should change the
content view controller of the root vc.

The changes in this commit are very rough and will be fleshed out later,
but this lends some initial structure to the future side menu view hierarchy.