Work on side menu
Set up a rudimentary side menu consisting of a table view with two
choices: Home and Start Game, which lead to the home vc and the start
game vc, respectively. The code is inefficient because each time the
user taps on a selection a new vc is instantiated (see
tableView:didSelectRowAtIndexPath: in TAPSideMenuViewController.m for
details).

Docs coming soon(tm).