Several changes to HomeVC -> now SummonerVC
SummonerVC is now a subclass of UITableViewController.
The summoner header is now a header of the table view,
and scrolls along with the recent matches below.

Header has a "background" image that is the splash art
of the last played champion in ranked.

Summoner icon now has a white border of 2 pixel/points.
Summoner level label is now in white for readability.

FIX: status bar is hard to see as black, but
should be ok when navigation controller is added as another color.