New class TAPSearchField
TAPSearchField implements a special textField
used for searching summoners (as of now).
It has a button as its rightView that shows
the current search region as its title.

Tapping region button when no text shows a toolbar
with a Done button that just dismissed the picker.
If region button already has text, shows toolbar
with a Search button that calls delegate's textFieldShouldReturn.

FIX: currently does not switch from picker to keyboard
when field is tapped and keyboard is already up.