Modified more loading behavior on SummonerVC.
Now starts loading when user scrolls to where footer is,
instead of starting when last cell is about to show.

This prevents the seemingly infinite scroll without warning
caused when user doesn't scroll fast enough to see the footer.
Ex. If user scrolls slow so that he stops at the last cell,
    app loads 15 more matches and its hard to tell which was the
    last cell I was looking at (esp. if duplicate champions).