Reuse of TAPSearchField for TAPPickerTextField
TAPPickerTextField is a variant of TAPSearchField that shows a picker
view when selected. The user should be able to tap the text field and
select a choice from a picker view that pops up.

The code right now is not very DRY because it replicates a large amount
of the code from TAPPickerTextField. It may be possible to create a
superclass for both classes.